create a token to authorization

+++ controllers/feed.js
@@ -1,29 +1,31 @@
-const fs = require("fs");
-const path = require("path");
+const fs = require('fs');
+const path = require('path');
 
-const { validationResult } = require("express-validator/check");
+const { validationResult } = require('express-validator/check');
 
-const Post = require("../models/post");
-const post = require("../models/post");
+const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
   const perPage = 2;
   let totalItems;
   Post.find()
-    .countDocument()
-    .then((count) => {
+    .countDocuments()
+    .then(count => {
       totalItems = count;
       return Post.find()
         .skip((currentPage - 1) * perPage)
         .limit(perPage);
     })
-    .then((posts) => {
-      res
-        .status(200)
-        .json({ message: "Fethched posts successfully", posts: posts, totalItems: totalItems, });
+    .then(posts => {
+      res.status(200).json({
+        message: 'Fetched posts successfully.',
+        posts: posts,
+        totalItems: totalItems
+      });
     })
-    .catch((err) => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
@@ -34,33 +36,43 @@ exports.getPosts = (req, res, next) => {
 exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    const error = new Error("Validation failed, entered data is incorrect.");
+    const error = new Error('Validation failed, entered data is incorrect.');
     error.statusCode = 422;
-    throw err;
+    throw error;
   }
   if (!req.file) {
-    const error = new Error("No image proided.");
-    err.statusCode = 422;
+    const error = new Error('No image provided.');
+    error.statusCode = 422;
     throw error;
   }
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Maximilian" },
+    creator: req.userId
   });
   post
     .save()
-    .then((result) => {
+    .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
-        message: "Post created successfully!",
-        post: result,
+        message: 'Post created successfully!',
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
-    .catch((err) => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
@@ -70,16 +82,16 @@ exports.createPost = (req, res, next) => {
 
 exports.getPost = (req, res, next) => {
   const postId = req.params.postId;
-  Post.findBydId(postId)
-    .then((post) => {
+  Post.findById(postId)
+    .then(post => {
       if (!post) {
-        const error = new Error("Could not find post.");
+        const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw error;
       }
-      return status(200).json({ message: "Post fetched", post: post });
+      res.status(200).json({ message: 'Post fetched.', post: post });
     })
-    .catch((err) => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
@@ -89,6 +101,12 @@ exports.getPost = (req, res, next) => {
 
 exports.updatePost = (req, res, next) => {
   const postId = req.params.postId;
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    const error = new Error('Validation failed, entered data is incorrect.');
+    error.statusCode = 422;
+    throw error;
+  }
   const title = req.body.title;
   const content = req.body.content;
   let imageUrl = req.body.image;
@@ -96,18 +114,23 @@ exports.updatePost = (req, res, next) => {
     imageUrl = req.file.path;
   }
   if (!imageUrl) {
-    const error = new Error("No file picked");
+    const error = new Error('No file picked.');
     error.statusCode = 422;
     throw error;
   }
   Post.findById(postId)
-    .then((post) => {
+    .then(post => {
       if (!post) {
-        const error = new Error("Could not find post.");
+        const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw error;
       }
-      if (imageUrl != post.imageUrl) {
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+      if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
       post.title = title;
@@ -115,13 +138,10 @@ exports.updatePost = (req, res, next) => {
       post.content = content;
       return post.save();
     })
-    .then((result) => {
-      res.status(200).json({
-        message: "post updated",
-        post: result,
-      });
+    .then(result => {
+      res.status(200).json({ message: 'Post updated!', post: result });
     })
-    .catch((err) => {
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
@@ -132,23 +152,32 @@ exports.updatePost = (req, res, next) => {
 exports.deletePost = (req, res, next) => {
   const postId = req.params.postId;
   Post.findById(postId)
-    .then((post) => {
+    .then(post => {
       if (!post) {
-        const error = new Error("Could not find post.");
+        const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw error;
       }
-      //check login user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!');
+        error.statusCode = 403;
+        throw error;
+      }
+      // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
-    .then((result) => {
-      console.log(result);
-      res.status(200).json({
-        message: "Deleted Post.",
-      });
+    .then(result => {
+      return User.findById(req.userId);
     })
-    .catch((err) => {
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
+      res.status(200).json({ message: 'Deleted post.' });
+    })
+    .catch(err => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
@@ -156,7 +185,7 @@ exports.deletePost = (req, res, next) => {
     });
 };
 
-const clearImage = (filePath) => {
-  filePath = path.join(__dirname, "..", filePath);
-  fs.unlink(filePath, (err) => console.log(err));
+const clearImage = filePath => {
+  filePath = path.join(__dirname, '..', filePath);
+  fs.unlink(filePath, err => console.log(err));
 };
